splitting http integraion test into its own file
