+++ test/integration/BUILD
@@ -15,7 +15,7 @@ envoy_cc_test(
     name = "proto_integration_test",
     srcs = ["proto_integration_test.cc"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/http:header_map_lib",
         "//source/common/protobuf:utility_lib",
         "//test/test_common:network_utility_lib",
@@ -52,7 +52,7 @@ envoy_cc_test(
         "//test/config/integration:server_http2.json",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/buffer:buffer_lib",
         "//source/common/http:header_map_lib",
         "//test/mocks/http:http_mocks",
@@ -69,7 +69,7 @@ envoy_cc_test(
     ],
     data = ["//test/config/integration:server_http2_upstream.json"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/http:header_map_lib",
         "//test/test_common:utility_lib",
     ],
@@ -85,7 +85,7 @@ envoy_cc_test(
         "//test/proto:bookstore_proto_descriptor",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/grpc:codec_lib",
         "//source/common/http:header_map_lib",
         "//test/proto:bookstore_proto",
@@ -104,8 +104,22 @@ envoy_cc_test(
         "//test/config/integration:server_config_files",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//include/envoy/http:header_map_interface",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "http_integration_lib",
+    srcs = [
+        "http_integration.cc",
+    ],
+    hdrs = [
+        "http_integration.h",
+    ],
+    data = ["//test/common/runtime:filesystem_test_data"],
+    deps = [
+        ":integration_lib",
     ],
 )
 
@@ -198,7 +212,7 @@ envoy_cc_test(
         "//test/config/integration:server_config_files",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/http:header_map_lib",
         "//source/server/config/stats:statsd_lib",
         "//test/test_common:utility_lib",
@@ -227,7 +241,7 @@ envoy_cc_test(
     ],
     data = ["//test/config/integration:server_proxy_proto.json"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/buffer:buffer_lib",
         "//source/common/http:codec_client_lib",
     ],
@@ -238,7 +252,7 @@ envoy_cc_test(
     srcs = ["ratelimit_integration_test.cc"],
     data = ["//test/config/integration:server_ratelimit.json"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/buffer:zero_copy_input_stream_lib",
         "//source/common/grpc:codec_lib",
         "//source/common/grpc:common_lib",
@@ -257,7 +271,7 @@ envoy_cc_test(
         "//test/config/integration/certs",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/event:dispatcher_includes",
         "//source/common/event:dispatcher_lib",
         "//source/common/network:utility_lib",
@@ -297,7 +311,7 @@ envoy_cc_test(
     ],
     data = ["//test/config/integration:server_uds.json"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/event:dispatcher_includes",
         "//source/common/event:dispatcher_lib",
         "//source/common/http:codec_client_lib",
@@ -311,7 +325,7 @@ envoy_cc_test(
     srcs = ["xds_integration_test.cc"],
     data = ["//test/config/integration:server_xds_files"],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
     ],
 )
 
@@ -327,7 +341,7 @@ envoy_cc_test(
         "//test/config/integration/certs",
     ],
     deps = [
-        ":integration_lib",
+        ":http_integration_lib",
         "//source/common/http:header_map_lib",
         "//test/test_common:utility_lib",
     ],
