+++ test/integration/integration_admin_test.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "test/integration/integration.h"
+#include "test/integration/http_integration.h"
 
 #include "gtest/gtest.h"
