+++ test/integration/proxy_proto_integration_test.h
@@ -4,7 +4,7 @@
 #include "common/stats/stats_impl.h"
 
 #include "test/integration/fake_upstream.h"
-#include "test/integration/integration.h"
+#include "test/integration/http_integration.h"
 #include "test/integration/server.h"
 
 #include "gtest/gtest.h"
