+++ test/integration/uds_integration_test.h
@@ -4,7 +4,7 @@
 #include "common/stats/stats_impl.h"
 
 #include "test/integration/fake_upstream.h"
-#include "test/integration/integration.h"
+#include "test/integration/http_integration.h"
 #include "test/integration/server.h"
 #include "test/test_common/environment.h"
