+++ test/integration/xds_integration_test.cc
@@ -1,4 +1,4 @@
-#include "test/integration/integration.h"
+#include "test/integration/http_integration.h"
 
 #include "gtest/gtest.h"
